✅ Rename Kotlin sample test of `Zero.Companion.fromDoubleOrNull`
This commit renames the Kotlin sample's test of the `Zero.Companion.fromDoubleOrNull(Double)` function.

Ref: #646

+++ subprojects/samples/src/test/kotlin/org/kotools/types/ZeroCompanionKotlinSampleTest.kt
@@ -44,6 +44,6 @@ class ZeroCompanionKotlinSampleTest {
         assertPrintsTrue(ZeroCompanionKotlinSample::fromFloatOrNull)
 
     @Test
-    fun `fromDoubleOrNull(Float) should pass`(): Unit =
+    fun `fromDoubleOrNull(Double) should pass`(): Unit =
         assertPrintsTrue(ZeroCompanionKotlinSample::fromDoubleOrNull)
 }
